Create requests in instance funcs and vars
This will form the basis of the revamped API. `Request` will gain a
phantom type of the returned object. These methods will then be made
public—and the methods on `Client` will be removed in favor of a single
method that executes a request.

`User` requests still need to migrated as they don’t map cleanly onto
the current data structure.